+++ crates/wolf_engine_window/src/lib.rs
@@ -94,11 +94,23 @@ mod window_system_tests {
             .with_title("Custom Test Title")
             .with_size((123, 123))
             .with_resizable(false)
-            .with_visible(false);
+            .with_visible(false)
+            .with_fullscreen_mode(FullscreenMode::Borderless);
 
         assert_eq!(window_settings.title, "Custom Test Title");
         assert_eq!(window_settings.size, (123, 123));
         assert_eq!(window_settings.is_resizable, false);
         assert_eq!(window_settings.is_visible, false);
+        assert_eq!(
+            window_settings.fullscreen_mode,
+            Some(FullscreenMode::Borderless)
+        );
+    }
+
+    #[test]
+    fn should_not_be_fullscreen_by_default() {
+        let window_settings = WindowSettings::default();
+
+        assert!(window_settings.fullscreen_mode.is_none());
     }
 }